Bug 1219530 - Add MOZ_RUST_MP4PARSE feature switch. r=glandium
Add a switch to enable the rust mp4parser code through confvars.sh
and set this for browser targets. Configure will only pass this
through as a CPP define if the rust toolchain is available.

The MOZ_RUST check is hoisted to an outer conditional to
make it cleaner to add other features.

Thanks to zhoubcfan163.com for the typo fix in configure.in.

UltraBlame original commit: 43ce9581fb647fe22dd5fa12069409cdd701b8a7