Bug 978833 patch 7 - Fuse allocation of ImportantStyleData with Declaration.  r=heycam
Note that this adds a new public API to css::Declaration; the equivalent
API is removed from css::StyleRule and nsCSSPageRule in patch 13.  But
the removal and addition need to be on opposite sides of patch 12.

This fused allocation is no larger than having a pointer, and it removes
having to worry about cycles.

UltraBlame original commit: 4b633979383a8174c450ed7f77c0f06b6095f404