Bug 978833 patch 16 - Always call Declaration::SetImmutable when we match a rule.  r=heycam
This change needs to happen before future work that would get rid of the
DeclarationChanged dance in which we make a new StyleRule, but I've
postponed that work to a later bug.  Without this, those changes would
cause a regression, because we'd only call SetImmutable on a StyleRule's
first mDeclaration.  However, we may as well do this now, as it makes
patch 17 and patch 18 possible.

UltraBlame original commit: 888d7ce53c3e032f6c4f6a70675a8775d710325d