Bug 1219711 - Let fake stream take precedence in testing. r=jib TL;DR requesting a fake stream always gives you a fake stream. No magic.
The gUMConstraint `fake: true` should take precedence and if set always
use MediaEngineDefault.
If it is set the state of `faketracks` is passed
on to MediaEngineDefault.
If it is not set, but (any of) audio/video loopback devices are set, the
device enumeration will filter out only those.

UltraBlame original commit: 4eaa8cf9a75277ae5b2c20586b2b2869a1db9a5e