Bug 1216030 - Part 9: We don't need to call CanPerformOnCompositorThread in RequestRestyle.
That's because KeyFrameEffectReadOnly::mIsPropertyRunningOnCompositor indicates the
correct state that the animation property is now running on compositor or not.
So once the flag is set to true, we do not need to check animation properties can run on
compositor after Animation::CanThrottle again.

UltraBlame original commit: 4158183a67ef086fdfd4a16e0226ef48490fa694