Bug 1222569 - fix initialization order in SourceSurfaceD2D1; r=Bas
Member fields are supposed to be initialized in the order they are
declared, but the constructor of SourceSurfaceD2D1 initialized
mDrawTarget prior to mDevice.  This is probably harmless, but it does
cause a warning on clang-cl, so let's fix it.

UltraBlame original commit: 5f9ff9850912fd3851e8ca41072ee250f45c876d