Bug 1222569 - fix initialization order in DataTextureSourceD3D9; r=Bas
Member fields are supposed to be initialized in the order they are
declared, but the constructors of DataTextureSourceD3D9 initialized
mFormat prior to mCompositor.  This is probably harmless, but it does
cause a warning on clang-cl, so let's fix it.

UltraBlame original commit: dff6ad030963f9ce971b57c139b75a9607b175f3