Bug 1219711 - Let fake stream take precedence in testing. r=jib TL;DR requesting a fake stream always gives you a fake stream. No magic.
The gUMConstraint `fake: true` should take precedence and if set always
use MediaEngineDefault.
If it is set the state of `faketracks` is passed
on to MediaEngineDefault.
If it is not set, but (any of) audio/video loopback devices are set, the
device enumeration will filter out only those.

UltraBlame original commit: 7b8755f0ccf56966d17e47bbabd6d35deea41cbc