Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
========

https://hg.mozilla.org/integration/gaia-central/rev/4deeea4a4e20
Author: Kevin Grandon <kevingrandonyahoo.com>
Desc: Merge pull request #33147 from Cwiiis/bug1220311-verticalhome-webapps-manage-for-iac

Bug 1220311 - Use webapps-manage so that IAC works correctly. r=kgrandon

========

https://hg.mozilla.org/integration/gaia-central/rev/1d2450faceb4
Author: Chris Lord <chrislord.netgmail.com>
Desc: Bug 1220311 - Use webapps-manage so that IAC works correctly. r=kgrandon

When handling an IAC message, the home screen will be launched as a
regular app and not as a home screen. Because of this, the
homescreen-webapps-manage permission isn't granted and the app fails to
launch with an exception caused by this. Revert to using the old
webapps-manage permission for now.

UltraBlame original commit: 93d2777599e9084abd4c41364d52e147dbc27322