Bug 1224046 - Remove <base href> from directory listings. r=mcmanus
There was an attempt to remove it in bug 358128, but that was backed out
because of FTP listings, where e.g. a 'mozilla' link on the
ftp://ftp.mozilla.org/pub listing (without a trailing slash) would go to
ftp://ftp.mozilla.org/pubmozilla instead of ftp://ftp.mozilla.org/pub/mozilla.

It appears that versions of Firefox at the time of that bug wouldn't "redirect"
directories to the same url with a trailing slash when one was not provided.
It also appears that current versions of Firefox *do* "redirect" in such a way,
such that the <base href> is now unnecessary.

That <base href> prevents links from resource://... directory listings to work
properly because of security contraints added in bug 1184387, as they force
links to go to a jar/file url instead of the equivalent resource://... url,
which removing <base href> allows.

UltraBlame original commit: 19e419ab3792fa5aad9f531364e72827ea01d34c