Bug 1226311, part 2 - Remove SpecialPowers.isMainProcess usage from various IndexedDB tests. r=janv
isMainProcess() is used in a few places to disable tests when run in a
content process. However, these tests are either never run in e10s or
pass when run with e10s, so remove all of the testing code. Some of
this was to work around issues with test_ipc.html which does not seem
to exist any more.

This patch also removes some code for disabling a test on Windows 2003.

UltraBlame original commit: f5d820b790a01bd27d94e23d5cf891ca10e71801