Bug 1189336 - (Part 2) Rename Restriction class to Restrictable. r=ally
This is the first step to move from a list of DISALLOW_* things to a list of
features that can be enabled or disabled for a profile. At a later stage we
will check whether a "restrictable" feature is allowed or not.

UltraBlame original commit: 82cd040777655c08b79aca030a9bd885c8fd9d88