Bug 1226440 - Expose a method to get a node's immediate dominator; r=bz,sfink
This commit adds the `getImmediateDominator` method to `DominatorTree` which
returns the id of the immediate dominator of the node associated with the given
id. This enables walking the dominator tree from leaves up parents all the way
to the root of the tree.

UltraBlame original commit: a28821163fff5d0db7a5015e0eeb891b30bb38b7